changes the component SnapshotFirebase.tsx into a util file FirebaseSnapshots and added a useEffect into the home page
