+++ README.md
@@ -181,5 +181,5 @@ const [buttonVariants, setButtonVariants] = useState<boolean[]>(
 -Database has been created inside Firebase
 -Firebase.ts has been updated to import Firestore
 -Data from Firestore is stored as a variable called filteredData as an array
--Exported component "SnapshotFirebase" that handles collecting data from Firebase using UseEffect
+-Exported util file "FirebaseSnapshots" that handles collecting data from Firebase in an async function
 -Data in the database contains fields: category, date, description, id, images and a name
